+++ nflow-engine/src/test/java/io/nflow/engine/NflowEngineTest.java
@@ -23,7 +23,6 @@ import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
 import io.nflow.engine.config.db.H2DatabaseConfiguration;
 import io.nflow.engine.service.DummyTestWorkflow;
 import io.nflow.engine.workflow.definition.AbstractWorkflowDefinition;
-import io.nflow.engine.workflow.definition.WorkflowState;
 import io.nflow.engine.workflow.executor.WorkflowExecutor;
 import io.nflow.engine.workflow.instance.QueryWorkflowInstances;
 import io.nflow.engine.workflow.instance.WorkflowInstance;
@@ -49,7 +48,7 @@ public class NflowEngineTest {
 
   @Test
   public void test() throws InterruptedException {
-    Collection<AbstractWorkflowDefinition<? extends WorkflowState>> workflowDefinitions = asList(new DummyTestWorkflow());
+    Collection<AbstractWorkflowDefinition> workflowDefinitions = asList(new DummyTestWorkflow());
     try (NflowEngine nflowEngine = new NflowEngine(dataSource(), new H2DatabaseConfiguration.H2SQLVariants(),
         workflowDefinitions)) {
       String type = "dummy";