+++ nflow-tests/src/test/java/io/nflow/tests/SignalWorkflowTest.java
@@ -1,6 +1,9 @@
 package io.nflow.tests;
 
+import static io.nflow.tests.demo.workflow.SlowWorkflow.INTERRUPTED;
+import static io.nflow.tests.demo.workflow.SlowWorkflow.SIGNAL_INTERRUPT;
 import static io.nflow.tests.demo.workflow.SlowWorkflow.SLOW_WORKFLOW_TYPE;
+import static io.nflow.tests.demo.workflow.TestState.PROCESS;
 import static java.lang.Thread.sleep;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
@@ -20,7 +23,6 @@ import io.nflow.rest.v1.msg.CreateWorkflowInstanceRequest;
 import io.nflow.rest.v1.msg.CreateWorkflowInstanceResponse;
 import io.nflow.rest.v1.msg.ListWorkflowInstanceResponse;
 import io.nflow.tests.demo.workflow.DemoWorkflow;
-import io.nflow.tests.demo.workflow.SlowWorkflow;
 import io.nflow.tests.extension.NflowServerConfig;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
@@ -54,18 +56,18 @@ public class SignalWorkflowTest extends AbstractNflowTest {
   public void checkSlowWorkflowIsRunning() throws Exception {
     for (int i = 0; i < 10; i++) {
       ListWorkflowInstanceResponse wf = getWorkflowInstance(resp.id);
-      if (wf != null && SlowWorkflow.State.process.name().equals(wf.state)) {
+      if (wf != null && PROCESS.name().equals(wf.state)) {
         return;
       }
       sleep(500);
     }
-    fail("Workflow did not enter state " + SlowWorkflow.State.process.name());
+    fail("Workflow did not enter state " + PROCESS.name());
   }
 
   @Test
   @Order(3)
   public void interruptWorkflowWithSignal() {
-    assertTrue(setSignal(resp.id, SlowWorkflow.SIGNAL_INTERRUPT, "Setting signal via REST API").setSignalSuccess);
+    assertTrue(setSignal(resp.id, SIGNAL_INTERRUPT, "Setting signal via REST API").setSignalSuccess);
   }
 
   @Test
@@ -73,7 +75,7 @@ public class SignalWorkflowTest extends AbstractNflowTest {
   public void checkSlowWorkflowIsInterrupted() throws Exception {
     for (int i = 0; i < 20; i++) {
       ListWorkflowInstanceResponse wf = getWorkflowInstance(resp.id);
-      if (SlowWorkflow.State.interrupted.name().equals(wf.state)) {
+      if (INTERRUPTED.name().equals(wf.state)) {
         return;
       }
       sleep(500);
@@ -88,16 +90,15 @@ public class SignalWorkflowTest extends AbstractNflowTest {
     assertThat(wf.actions.size(), is(4));
     Action action = wf.actions.get(0);
     assertThat(action.stateText, is("Interrupted with signal 1, moving to interrupted state"));
-    assertThat(action.state, is(SlowWorkflow.State.process.name()));
+    assertThat(action.state, is(PROCESS.name()));
     assertThat(action.type, is(WorkflowActionType.stateExecution.name()));
     action = wf.actions.get(1);
     assertThat(action.stateText, is("Clearing signal from process state"));
-    assertThat(action.state, is(SlowWorkflow.State.process.name()));
+    assertThat(action.state, is(PROCESS.name()));
     assertThat(action.type, is(WorkflowActionType.stateExecution.name()));
     action = wf.actions.get(2);
     assertThat(action.stateText, is("Setting signal via REST API"));
-    assertThat(action.state, is(SlowWorkflow.State.process.name()));
+    assertThat(action.state, is(PROCESS.name()));
     assertThat(action.type, is(WorkflowActionType.externalChange.name()));
   }
-
 }