+++ CHANGELOG.md
@@ -2,15 +2,21 @@
 
 **Highlights**
 - `nflow-engine`
+  - BREAKING CHANGE: Remove `WorkflowDefinition`, workflow definitions should extend `AbstractWorkflowDefinition` instead
+  - BREAKING CHANGE: Remove deprecated `WorkflowState.isRetryAllowed`, set exception analyzer for workflow definition instead (if needed)
   - Enable maintenance (archiving and deleting old workflow instances) by default
   - Enable workflow instance history clean-up (deleting old actions and state variables) by default
 
 **Details**
 - `nflow-engine`
+  - Workflow definitions that used to extend `WorkflowDefinition` should now extend `AbstractWorkflowDefinition` instead.
+    - It is not necessary to define the workflow states as an enum anymore, which should make it easier to extend and reuse states across different workflow definitions.
+    - You can define the states as instances of `io.nflow.engine.workflow.curated.State` or anything else that implements the required `io.nflow.engine.workflow.definition.WorkflowState` interface.
+    - The workflow definitions must now register all possible states as described in `io.nflow.engine.workflow.definition.AbstractWorkflowDefinition`.
+  - `WorkflowState.isRetryAllowed` was removed. If it was overridden, you can use `new WorkflowSettings.Builder().setExceptionAnalyzer(...)` to change the behavior. The default behavior was not changed.
   - Maintenance workflow instance is added to nFlow database by default in startup. Instances that have been in final state longer than 45 days are archived. Archived instances that have been in final state longer than one year are deleted. Maintenance is run every night. Use `nflow.maintenance` configuration options to change the defaults before startup, or update the maintenance workflow instance state variables after the instance has been created.
   - Workflow instance actions and state variables that are older than 45 days are automatically cleaned up occasionally when the instance is processed. Use workflow settings to change the default time period (`setHistoryDeletableAfter`) and condition (`setDeleteHistoryCondition`) of the clean-up.
-  - Dependency updates:
-    - minimum supported maven version for buliding is 3.8
+  - Dependency updates
     - logback-classic update to version 1.2.10
       - http://mailman.qos.ch/pipermail/announce/2021/000164.html
       - https://jira.qos.ch/browse/LOGBACK-1591
@@ -24,6 +30,7 @@
     - jodatime 2.10.3
     - slf4j 1.7.32
 - `nflow-rest-api`
+  - Dependency updates
     - swagger 1.6.4
 - `nflow-jetty`
   - Dependency updates
@@ -36,12 +43,14 @@
   - Dependency updates
     - metrics 4.2.7
 - `nflow-tests`
+  - Dependency updates
     - h2 2.0.206
-      - Note: If you have persisted any h2 databases you must take a backup and restore. Also the nflow h2 schema changed to work with 2.x release of h2.
+      - Note: If you have persisted any h2 databases you must take a backup and restore. Also the nFlow h2 schema changed to work with 2.x release of h2.
     - mssql 9.4.1
     - mysql 8.0.27
     - mariadb 2.7.4
     - postgresql 42.3.1
+- Minimum supported Maven version for building is 3.6
 
 ## 7.4.0 (2021-12-27)